# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZiloPay API",
        "description": "Payment Processing and Money Transfer API",
        "contact": {
            "name": "ZiloPay Support",
            "email": "support@zilopay.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/merchant": {
            "get": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Get merchant by UUID",
                "operationId": "19815424c427f87fb840f8aefc4566ae",
                "responses": {
                    "200": {
                        "description": "Merchant retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Merchant retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Merchant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Merchant not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Merchants"
                ],
                "summary": "Create merchant profile",
                "operationId": "a2c5a19a96cdc2e83b90bbb65142e740",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "business_name",
                                    "business_email",
                                    "country"
                                ],
                                "properties": {
                                    "business_name": {
                                        "type": "string",
                                        "example": "My Shop"
                                    },
                                    "business_email": {
                                        "type": "string",
                                        "example": "contact@myshop.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+22501234567"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "CI"
                                    },
                                    "documents[id_card]": {
                                        "description": "ID card PDF file (max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "documents[business_license]": {
                                        "description": "Business license PDF file (max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "documents[tax_certificate]": {
                                        "description": "Tax certificate PDF file (max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Merchant profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Merchant profile created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Merchant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Merchant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "uuid": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "business_name": {
                        "type": "string",
                        "example": "ZiloShop"
                    },
                    "business_email": {
                        "type": "string",
                        "example": "contact@ziloshop.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+22501234567"
                    },
                    "country": {
                        "type": "string",
                        "example": "CI"
                    },
                    "fee_fixed": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "fee_percent": {
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "status_label": {
                        "type": "string",
                        "example": "Pending"
                    },
                    "approved_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15 10:30:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15 10:30:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based auth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}